+++ user/models.py
@@ -1,0 +1,11 @@
+from django.contrib.auth.models import AbstractUser
+from django.db import models
+
+# Create your models here.
+class User(AbstractUser):
+    phone_number = models.CharField(max_length=15, unique=True, null=True, blank=True)
+    is_phone_verified = models.BooleanField(default=False)
+
+    def __str__(self):
+        return self.username
+
